requirements.txt
streamlit
google-generativeai
plotly
pandas
python-dotenv
datetime
json5
typing-extensions